bump 20.09.03
- Added Multi-Network for Multisite.
- Added Object OPcache, WP OPcache stats.
- Added lookup_* methods to handle our temp internal data.
- Added locking file to suspend cache addition when doing flush.
- Fixed replace update_user_meta with lookup function to makes query-monitor happy.
- Fixed Admin interface, loading spinner should not display when no action.
- Fixed CronAgent::run_wpcron(), reset doing_cron if locked.
- Fixed CronAgent::run_wpcron(), halt if run reach maximum CRONBOX_MAX for site in multisite.

+++ includes/src/Dropino.php
@@ -15,10 +15,12 @@ namespace Nawawi\DocketCache;
 final class Dropino extends Bepart
 {
     private $path;
+    private $wpcondir;
 
     public function __construct($path)
     {
         $this->path = $path;
+        $this->wpcondir = \defined('WP_CONTENT_DIR') ? WP_CONTENT_DIR : ABSPATH.'wp-content';
     }
 
     /**
@@ -28,11 +30,11 @@ final class Dropino extends Bepart
     {
         $dt = [];
         $dt['src'] = $this->path.'/includes/object-cache.php';
-        $dt['dst'] = WP_CONTENT_DIR.'/object-cache.php';
+        $dt['dst'] = $this->wpcondir.'/object-cache.php';
 
         // sync with includes/object-cache.php
-        $dt['halt'] = WP_CONTENT_DIR.'/.object-cache-delay.txt';
-        $dt['after'] = WP_CONTENT_DIR.'/.object-cache-after-delay.txt';
+        $dt['halt'] = $this->wpcondir.'/.object-cache-delay.txt';
+        $dt['after'] = $this->wpcondir.'/.object-cache-after-delay.txt';
 
         return (object) $dt;
     }
@@ -42,7 +44,7 @@ final class Dropino extends Bepart
      */
     public function exists()
     {
-        return @is_file(WP_CONTENT_DIR.'/object-cache.php');
+        return @is_file($this->wpcondir.'/object-cache.php');
     }
 
     /**
@@ -56,7 +58,7 @@ final class Dropino extends Bepart
             return $cache[$key];
         }
 
-        $cache['dropino'] = $this->plugin_meta(WP_CONTENT_DIR.'/object-cache.php');
+        $cache['dropino'] = $this->plugin_meta($this->wpcondir.'/object-cache.php');
         $cache['plugin'] = $this->plugin_meta($this->path.'/includes/object-cache.php');
 
         return $cache[$key];
@@ -153,7 +155,11 @@ final class Dropino extends Bepart
                 $this->delay();
             }
 
-            return $this->copy($src, $dst);
+            if ($this->copy($src, $dst)) {
+                $this->multinet_active(true);
+
+                return true;
+            }
         }
 
         return false;
@@ -172,6 +178,14 @@ final class Dropino extends Bepart
             return true;
         }
 
+        // remove flag file to trigger disable at admin interface
+        $this->multinet_active(false);
+
+        // dont remove drop-in if active on sub network.
+        if ((!\defined('WP_CLI') || !WP_CLI) && $this->multinet_available()) {
+            return true;
+        }
+
         $this->opcache_flush($file);
 
         if (is_writable($file) && @unlink($file)) {
@@ -179,5 +193,181 @@ final class Dropino extends Bepart
         }
 
         return false;
+    }
+
+    public function multinet_install($hook)
+    {
+        if (!is_multisite() || !nwdcx_network_multi()) {
+            return false;
+        }
+
+        if (!nwdcx_wpdb($wpdb)) {
+            return false;
+        }
+
+        $table = $wpdb->base_prefix.'sitemeta';
+
+        $network_id = null;
+
+        $suppress = $wpdb->suppress_errors(true);
+
+        $query = "SELECT `site_id`,`meta_key`,`meta_value` FROM `{$table}` WHERE `meta_key`='active_sitewide_plugins' ORDER BY site_id ASC LIMIT 100";
+        $results = $wpdb->get_results($query, ARRAY_A);
+        if (!empty($results) && \is_array($results)) {
+            while ($row = @array_shift($results)) {
+                $site_id = $row['site_id'];
+                $meta_key = $row['meta_key'];
+                $meta_value = maybe_unserialize($row['meta_value']);
+
+                if (null === $network_id) {
+                    $network_id = $site_id;
+                }
+
+                if (!empty($meta_value) && \is_array($meta_value) && !empty($meta_value[$hook])) {
+                    $this->multinet_active(true, $site_id);
+                }
+            }
+        }
+
+        if (null !== $network_id) {
+            @file_put_contents($this->wpcondir.'/.object-cache-network-main.txt', $network_id, FILE_EX);
+        }
+
+        $wpdb->suppress_errors($suppress);
+    }
+
+    private function multinet_clear_main($cleanup = false)
+    {
+        $file = $this->wpcondir.'/.object-cache-network-main.txt';
+        if (@is_file($file)) {
+            @unlink($file);
+        }
+
+        if ($cleanup) {
+            $file = $this->wpcondir.'/.object-cache-network-multi.txt';
+            if (@is_file($file)) {
+                @unlink($file);
+            }
+        }
+    }
+
+    private function multinet_list()
+    {
+        $list = [];
+        $files = @glob($this->wpcondir.'/.object-cache-network-*.txt', GLOB_MARK | GLOB_NOSORT);
+        if (!empty($files) && \is_array($files)) {
+            foreach ($files as $file) {
+                $fx = basename($file);
+                if (@preg_match('@^\.object-cache-network-(\d+)\.txt$@', $fx, $mm)) {
+                    $id = $mm[1];
+                    $list[$id] = $file;
+                }
+            }
+        }
+
+        return !empty($list) ? $list : false;
+    }
+
+    public function multinet_available()
+    {
+        $network_id = get_current_network_id();
+        $files = $this->multinet_list();
+        if (!empty($files) && \is_array($files)) {
+            foreach ($files as $id => $file) {
+                if ($id !== $network_id) {
+                    return true;
+                }
+            }
+        }
+
+        return false;
+    }
+
+    public function multinet_clear($cache_path, $logfile)
+    {
+        clearstatcache();
+        $this->multinet_clear_main(true);
+
+        $files = $this->multinet_list();
+        if (!empty($files) && \is_array($files)) {
+            foreach ($files as $network_id => $file) {
+                $cachepath = $cache_path;
+                if (@is_file($file)) {
+                    if (false === strpos($cache_path, '/network-')) {
+                        $cachepath = $cache_path.'/network-'.$network_id.'/';
+                    }
+
+                    if (@is_dir($cachepath)) {
+                        $this->cachedir_flush($cachepath);
+                    }
+                    @unlink($file);
+                }
+
+                $ext = substr($logfile, -4);
+                $fname = substr($logfile, 0, -4);
+                $logfile = $fname.'-'.$network_id.$ext;
+                if (@is_file($logfile)) {
+                    @unlink($logfile);
+                }
+            }
+        }
+    }
+
+    public function multinet_tag($network_id = false)
+    {
+        $network_id = empty($network_id) ? get_current_network_id() : $network_id;
+
+        return sprintf('%s/.object-cache-network-%s.txt', $this->wpcondir, $network_id);
+    }
+
+    public function multinet_active($status = false, $network_id = false)
+    {
+        if (!is_multisite() || !nwdcx_network_multi()) {
+            return false;
+        }
+
+        $lock_file = $this->wpcondir.'/.object-cache-network-multi.txt';
+        if (!@is_file($lock_file)) {
+            @file_put_contents($lock_file, 1);
+        }
+
+        if (empty($network_id)) {
+            $this->multinet_clear_main();
+            $network_id = get_current_network_id();
+        }
+
+        $file = $this->multinet_tag($network_id);
+
+        clearstatcache();
+
+        $is_file = @is_file($file);
+        if ($status) {
+            if ($is_file) {
+                return true;
+            }
+
+            return $this->put($file, $network_id);
+        }
+
+        if (!$is_file) {
+            return true;
+        }
+
+        return @unlink($file);
+    }
+
+    public function multinet_me()
+    {
+        if (!is_multisite()) {
+            return true;
+        }
+
+        if (!nwdcx_network_multi()) {
+            return true;
+        }
+
+        $file = $this->multinet_tag();
+
+        return @is_file($file);
     }
 }
